Refactoring elements
