+++ pygame_gui/core/components/__init__.py
@@ -1,0 +1,5 @@
+from pygame_gui.core.components.component import Component
+from pygame_gui.core.components.transform import Transform
+
+__all__ = ['Component',
+           'Transform']
