+++ pygame_gui/elements/ui_button.py
@@ -78,13 +78,15 @@ class UIButton(UIElement):
                          parent_element=parent_element,
                          object_id=object_id,
                          element_id=['button'])
+        
+        self.transform.dynamic_dimensions_orig_top_left = rel_rect.topleft
+        self.transform.dynamic_size_calculator = lambda: self.image.get_size()
 
         self.text = text
         self.text_kwargs = {}
         if text_kwargs is not None:
             self.text_kwargs = text_kwargs
 
-        self.dynamic_dimensions_orig_top_left = rel_rect.topleft
         # support for an optional 'tool tip' element attached to this button
         self.set_tooltip(tool_tip_text, tool_tip_object_id, tool_tip_text_kwargs)
         self.ui_root_container = self.ui_manager.get_root_container()
@@ -772,25 +774,6 @@ class UIButton(UIElement):
         self.on_fresh_drawable_shape_ready()
 
         self._on_contents_changed()
-
-    def _calc_dynamic_size(self):
-        if self.dynamic_width or self.dynamic_height:
-            self._set_dimensions(self.image.get_size())
-
-            # if we have anchored the left side of our button to the right of its container then
-            # changing the width is going to mess up the horiz position as well.
-            new_left = self.relative_rect.left
-            new_top = self.relative_rect.top
-            if 'left' in self.anchors and self.anchors['left'] == 'right' and self.dynamic_width:
-                left_offset = self.dynamic_dimensions_orig_top_left[0]
-                new_left = left_offset - self.relative_rect.width
-            # if we have anchored the top side of our button to the bottom of it's container then
-            # changing the height is going to mess up the vert position as well.
-            if 'top' in self.anchors and self.anchors['top'] == 'bottom' and self.dynamic_height:
-                top_offset = self.dynamic_dimensions_orig_top_left[1]
-                new_top = top_offset - self.relative_rect.height
-
-            self.set_relative_position((new_left, new_top))
 
     def hide(self):
         """
